requirements.txt
Django==6.0.4
django-jazzmin==3.0.1
aiogram==3.20.0.post0
python-dotenv==1.1.0
psycopg[binary]==3.2.6
Pillow==11.2.1
django-environ==0.12.0
whitenoise==6.9.0
gunicorn==23.0.0
openpyxl==3.1.5
